added class Map and change way to get data
