+++ src/psi_environment/api/environment_api.py
@@ -1,25 +1,25 @@
 import numpy as np
 
-from src.psi_environment.environment import Environment
+from psi_environment.data.map import Map
 
 
 class EnvironmentAPI:
-    def __init__(self, environment: Environment):
-        self.environment = environment
+    def __init__(self, _map: Map):
+        self._map = _map
 
     def get_adjacency_matrix(self):
         """
         Returns the adjacency matrix of the environment's nodes
         :return: numpy array with shape (num_nodes, num_nodes)
         """
-        return self.environment.get_map()
+        return self._map.get_adjacency_matrix()
 
     def get_traffic(self):
         """
         Returns the traffic matrix
         :return: numpy array with shape (num_nodes, num_nodes)
         """
-        size = self.environment.get_map_size()
+        size = self._map.get_adjacency_matrix_size()
         traffic_matrix = np.full((size, size), np.nan)
 
         for i in range(size):
@@ -33,12 +33,12 @@ class EnvironmentAPI:
         :param to_node: integer representing the index of end node
         :return: integer represents traffic from node to node
         """
-        size = self.environment.get_map_size()
+        size = self._map.get_adjacency_matrix_size()
 
         if 0 > from_node > size and 0 > to_node > size:
             raise ValueError("Incorrect values of from_node and to_node")
 
-        edges = self.environment.get_edges()
+        edges = self._map.get_edges()
 
         if (from_node, to_node) not in edges.keys():
             return np.nan
