+++ src/psi_environment/environment.py
@@ -1,46 +1,9 @@
 import random
 from typing import Any
 
-import numpy as np
-from collections import deque
-
 from psi_environment.game.game import Game
-
-
-def get_map(filename="./game/resources/sample_map.txt") -> np.ndarray:
-    """
-    Generates adjacency matrix of a sample map saved in ./game/resources/sample.map.txt
-    :return: adjacency matrix of the map as a numpy array
-    """
-    with open(filename) as f:
-        content = f.read()
-
-    content = content.split()
-    connecting_characters = {"=", "x"}
-    content = np.array([[*row] for row in content])
-
-    node_indices = {}
-    index = 0
-    for y in range(content.shape[0]):
-        for x in range(content.shape[1]):
-            if content[y][x] in connecting_characters:
-                node_indices[(x, y)] = index
-                index += 1
-
-    n_nodes = len(node_indices)
-
-    adjacency_matrix = np.full((n_nodes, n_nodes), np.nan)
-
-    moves = np.array([[1, 0], [0, 1], [-1, 0], [0, -1]])
-
-    for (x, y), node_index in node_indices.items():
-        for dx, dy in moves:
-            nx, ny = x + dx, y + dy
-            if (nx, ny) in node_indices and content[ny][nx] in connecting_characters:
-                neighbor_index = node_indices[(nx, ny)]
-                adjacency_matrix[node_index, neighbor_index] = 1
-
-    return adjacency_matrix
+from psi_environment.data.map import Map
+from psi_environment.api.environment_api import EnvironmentAPI
 
 
 class Environment:
@@ -48,16 +11,10 @@ class Environment:
         if random_seed is None:
             random_seed = random.randint(0, 2137)
         self._random_seed = random_seed
-        self._map = get_map()
+        self._map = Map()
+        self.data = EnvironmentAPI(self._map)
         self._game = Game(self._map, random_seed=random_seed)
         self._is_running = True
-        # Preparing 2-directional queue for each connection between crossroads
-        self._edges = {}
-        for y in range(self._map.shape[0]):
-            for x in range(y, self._map.shape[1]):
-                if self._map[y, x] != np.nan:
-                    self._edges[(y, x)] = deque()
-                    self._edges[(x, y)] = deque()
 
     def step(self, action: int) -> tuple[Any, float, bool]:
         """
@@ -86,15 +43,6 @@ class Environment:
         """
         return self._game.reset()
 
-    def get_map_size(self):
-        return self._map.shape[0]
-
-    def get_map(self):
-        return self._map
-
-    def get_edges(self):
-        return self._edges
-    
     def is_running(self):
         self._is_running = self._game.is_running()
         return self._is_running
